+++ src/Storage/TimeToLiveTest.php
@@ -21,39 +21,69 @@ final class TimeToLiveTest extends TestCase
     }
 
     /**
-     * @param object|string|int $input
      * @dataProvider validDurationString
      */
-    public function testItCanBeInstantiatedFromDurationInput($input, DateInterval $expected): void
+    public function testItCanBeInstantiatedFromDurationInput(string $input, DateInterval $expected): void
     {
         self::assertEquals($expected, TimeToLive::fromDurationString($input));
     }
 
     /**
-     * @return iterable<string, array{input:int|string|object, expected:DateInterval}>
+     * @return iterable<string, array{input:string, expected:DateInterval}>
      */
     public function validDurationString(): iterable
     {
         $threeDays = new DateInterval('P3D');
-
-        yield 'stringable object' => [
-            'input' => new class() {
-                public function __toString(): string
-                {
-                    return '3 days';
-                }
-            },
-            'expected' => $threeDays,
-        ];
 
         yield 'string' => [
             'input' => '3 days',
             'expected' => $threeDays,
         ];
 
+        yield 'stringable negative days' => [
+            'input' => '-3 days',
+            'expected' => DateInterval::createFromDateString('-3 DAYS'),
+        ];
+    }
+
+    /**
+     * @param int|string|object|null|DateInterval $input
+     * @param ?DateInterval                       $expected
+     * @dataProvider validDurationInt
+     */
+    public function testItCanBeInstantiatedFromSeconds($input, ?DateInterval $expected): void
+    {
+        self::assertEquals($expected, TimeToLive::convert($input));
+    }
+
+    /**
+     * @return iterable<string, array{input:int|string|object|null|DateInterval, expected:DateInterval|null}>
+     */
+    public function validDurationInt(): iterable
+    {
         $seconds = new DateInterval('PT2345S');
 
-        yield 'stringable seconds' => [
+        yield 'DateInterval' => [
+            'input' => $seconds,
+            'expected' => $seconds,
+        ];
+
+        yield 'null' => [
+            'input' => null,
+            'expected' => null,
+        ];
+
+        yield 'stringable object' => [
+            'input' => new class() {
+                public function __toString(): string
+                {
+                    return '2345';
+                }
+            },
+            'expected' => $seconds,
+        ];
+
+        yield 'numeric string' => [
             'input' => '2345',
             'expected' => $seconds,
         ];
@@ -63,12 +93,9 @@ final class TimeToLiveTest extends TestCase
             'expected' => $seconds,
         ];
 
-        $negativeInterval = clone $seconds;
-        $negativeInterval->invert = 1;
-
         yield 'negative seconds' => [
             'input' => '-2345',
-            'expected' => $negativeInterval,
+            'expected' => DateInterval::createFromDateString('-2345 seconds'),
         ];
     }
 }